+++ resources/dicts/events/death/child_dislike.json
