+++ resources/dicts/events/death/child_jealousy.json
