+++ resources/dicts/events/death/general_dislike.json
