+++ resources/dicts/events/death/general_jealousy.json
