+++ scripts/game_structure/game_essentials.py
@@ -1,7 +1,7 @@
 import pygame
 import pygame_gui
 
-from scripts.datadir import get_save_dir, get_temp_dir
+from scripts.housekeeping.datadir import get_save_dir, get_temp_dir
 
 import ujson
 import os
