+++ scripts/housekeeping/noop_writer.py
