+++ scripts/housekeeping/update.py
@@ -15,9 +15,9 @@ import requests
 from requests import Response
 from strenum import StrEnum
 
-from scripts.progress_bar_updater import UIUpdateProgressBar
+from scripts.housekeeping.progress_bar_updater import UIUpdateProgressBar
 from scripts.utility import quit
-from scripts.version import get_version_info
+from scripts.housekeeping.version import get_version_info
 
 use_proxy = False  # Set this to True if you want to use a proxy for the update check. Useful for debugging.
