+++ scripts/screens/organizational_screens.py
@@ -31,12 +31,12 @@ from scripts.game_structure.game_essentials import game, screen, screen_x, scree
 from scripts.game_structure.windows import DeleteCheck, UpdateAvailablePopup, ChangelogPopup, SaveError
 from scripts.game_structure.discord_rpc import _DiscordRPC
 from scripts.game_structure import image_cache
-from ..datadir import get_data_dir, get_cache_dir
-from ..update import has_update, UpdateChannel, get_latest_version_number
+from ..housekeeping.datadir import get_data_dir, get_cache_dir
+from ..housekeeping.update import has_update, UpdateChannel, get_latest_version_number
 
 import ujson
 
-from ..version import get_version_info
+from ..housekeeping.version import get_version_info
 
 logger = logging.getLogger(__name__)
 has_checked_for_update = False
