+++ tests/test_save.py
@@ -11,7 +11,7 @@ os.environ["SDL_AUDIODRIVER"] = "dummy"
 os.environ["SDL_VIDEODRIVER"] = "dummy"
 os.environ["SDL_AUDIODRIVER"] = "dummy"
 
-from scripts.datadir import get_save_dir
+from scripts.housekeeping.datadir import get_save_dir
 from scripts.game_structure.game_essentials import Game
 
 if not os.path.exists('tests/testSaves'):
